major refactoring - tests/formating to fix
