+++ trimesh/voxel/__init__.py
@@ -1,0 +1,12 @@
+from .base import Voxel
+from .base import VoxelBase
+from .base import VoxelMesh
+from .base import VoxelRle
+
+
+__all__ = [
+    'Voxel',
+    'VoxelBase',
+    'VoxelMesh',
+    'VoxelRle',
+]
