+++ trimesh/voxel/runlength.py
